pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "freelance-assitant"
version = "0.1.0"
description = "Personal freelance lead-generation system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "arq>=0.26",
    "aiogram>=3.10",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic-settings>=2.5",
    "click>=8.1",
    "feedparser>=6.0",
    "openai>=1.40",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "aiosqlite>=0.20",
    "ruff>=0.6",
]
browser = [
    "playwright>=1.45",
]
telegram-channels = [
    "telethon>=1.36",
]

[project.scripts]
fa = "freelance_assitant.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/freelance_assitant"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
